+++ R/get_pvol.R
@@ -10,6 +10,11 @@
 #' are operational. For example, radars with the `status` `0` in `get_weather_radars("opera")`
 #' are currently not operational.
 #'
+#' For European countries where a dedicated archive exist this data is used. Alternatively
+#' data is downloaded from the Opera open radar data archive. This data is frequently
+#' filtered. Use the argument `use_opera_ord = TRUE` if you want to force using the
+#' Opera open radar data.
+#'
 #' Not all radars in the nexrad archive can be read successfully. Radars associated
 #' with the Terminal Doppler Weather Radar (TDWR) program can not be read. These can
 #' be identified using the `stntype` column in `get_weather_radars("nexrad")`.
@@ -32,12 +37,19 @@
 #' # Get PVOL data for a single radar and datetime
 #' get_pvol("deess", as.POSIXct(Sys.Date()))
 #'
+#' # Get similar data from the opera open radar data
+#' get_pvol("deess", as.POSIXct(Sys.Date()), use_opera_ord=TRUE)
+#'
 #' # Get PVOL data for multiple radars and a single datetime
 #' get_pvol(
 #'   c("deess", "dehnr", "fianj", "czska", "KABR"),
 #'   as.POSIXct(Sys.Date())
 #' )
-get_pvol <- function(radar = NULL, datetime = NULL, ...) {
+get_pvol <- function(
+  radar = NULL,
+  datetime = NULL,
+  ...
+) {
   if (!identical(radar, "hochficht")) {
     check_odim_nexrad(setdiff(radar, 'hochficht'))
   }
@@ -82,7 +94,7 @@ get_pvol <- function(radar = NULL, datetime = NULL, ...) {
   # ensure early failure and not first do a lot of download before failing on
   # the last radar
   if (length(radar) != 1) {
-    purrr::map(radar, select_get_pvol_function) # quick check if all radars exist
+    purrr::map(radar, select_get_pvol_function, ...) # quick check if all radars exist
     pvols <- purrr::map(radar, safe_get_pvol, datetime = datetime, ...)
     if (lubridate::is.interval(datetime)) {
       pvols <- unlist(pvols, recursive = F)
@@ -90,7 +102,7 @@ get_pvol <- function(radar = NULL, datetime = NULL, ...) {
     return(pvols)
   }
 
-  fn <- select_get_pvol_function(radar)
+  fn <- select_get_pvol_function(radar, ...)
 
   if (lubridate::is.interval(datetime)) {
     if (lubridate::as.duration(datetime) > lubridate::hours(1)) {
@@ -136,16 +148,23 @@ get_pvol <- function(radar = NULL, datetime = NULL, ...) {
 
 # Helper function to find the function for a specific radar
 # This function is only helpful in get_pvol and therefor not in a utils file
-select_get_pvol_function <- function(radar, ..., call = rlang::caller_env()) {
+select_get_pvol_function <- function(
+  radar,
+  ...,
+  use_opera_ord = FALSE,
+  call = rlang::caller_env()
+) {
   if (radar == "hochficht") {
     return("get_pvol_hochficht")
   }
+
   if (is_nexrad(radar)) {
     return("get_pvol_us")
   }
 
   cntry_code <- substr(radar, 1, 2) # nolint
   fun <- (dplyr::case_when(
+    use_opera_ord ~ "get_pvol_ord",
     cntry_code == "nl" ~ "get_pvol_nl",
     cntry_code == "fi" ~ "get_pvol_fi",
     cntry_code == "dk" ~ "get_pvol_dk",
@@ -155,6 +174,30 @@ select_get_pvol_function <- function(radar, ..., call = rlang::caller_env()) {
     cntry_code == "se" ~ "get_pvol_se",
     cntry_code == "ro" ~ "get_pvol_ro",
     cntry_code == "sk" ~ "get_pvol_sk",
+    cntry_code %in%
+      c(
+        "be",
+        "ch",
+        "cz",
+        "de",
+        "dk",
+        "ee",
+        "es",
+        "fi",
+        "fr",
+        "hr",
+        "ie",
+        "is",
+        "lt",
+        "mt",
+        "nl",
+        "no",
+        "pl",
+        "ro",
+        "se",
+        "si"
+      ) ~
+      "get_pvol_ord", # Some countries here can't be reached as a custom function exists
     .default = NA
   ))
   if (rlang::is_na(fun)) {